+++ R/dbfmcl_main.R
@@ -32,17 +32,19 @@ print_msg <- function(msg, msg_type="INFO"){
 library(ggplot2)
 library(reshape2)
 library(dplyr)
-library(MCL)
 library(igraph)
+library(iheatmapr)
 
 #' @title
 #' ClusterSet
 #' @description
 #' This class is a representation of a partitioning algorithm and is intented to store gene clusters.
 #' @slot name character. The original input file name (if applicable).
+#' @slot opt_name character. The name is file containing distance and cutting threshold.
 #' @slot data matrix. The matrix containing the filtered/partitionned data.
 #' @slot cluster vector. Mapping of row/genes to clusters.
 #' @slot size vector. The size of each cluster.
+#' @slot top_genes matrix The highly co-expressed genes of each gene clusters.
 #' @slot center matrix. The centers of each clusters.
 #' @slot parameters list. The parameter used.
 #' @slot algorithm vector. The algorithm used to produce the clusters.
@@ -75,9 +77,11 @@ library(igraph)
 setClass("ClusterSet",
   representation = list(
     name = "character",
+    opt_name = "character",
     data = "matrix",
     cluster = "vector",
     size = "vector",
+    top_genes = "matrix",
     center = "matrix",
     parameters = "list",
     algorithm = "vector",
@@ -91,6 +95,7 @@ setClass("ClusterSet",
     data = matrix(nr = 0, nc = 0),
     cluster = numeric(),
     size = numeric(),
+    top_genes = matrix(),
     center = matrix(nc = 0, nr = 0),
     parameters = list(),
     algorithm = character(),
@@ -286,15 +291,15 @@ setMethod("load_seurat",
         stop("The number of cells is not the same in ClusterSet object and Seurat object.")
       }
 
-      g <- ggplot2::ggplot_build(dimplot_obj)
-      tmp_mat <- dplyr::distinct(as.data.frame(cbind(g$data[[1]]$colour,
+      g <- ggplot_build(dimplot_obj)
+      tmp_mat <- distinct(as.data.frame(cbind(g$data[[1]]$colour,
                                               as.character(g$plot$data$ident))))
       cell_col_tmp <- tmp_mat[,1]
       cell_grp_tmp <- tmp_mat[,2]
-      object@cell_colors <- stats::setNames(as.character(cell_col_tmp), cell_grp_tmp)
+      object@cell_colors <- setNames(as.character(cell_col_tmp), cell_grp_tmp)
 
-      tmp_mat <- dplyr::distinct(as.data.frame(cbind(rownames(g$plot$data), as.character(g$plot$data$ident))))
-      object@cell_types <- stats::setNames(as.character(tmp_mat[,2]), tmp_mat[,1])
+      tmp_mat <- distinct(as.data.frame(cbind(rownames(g$plot$data), as.character(g$plot$data$ident))))
+      object@cell_types <- setNames(as.character(tmp_mat[,2]), tmp_mat[,1])
 
       object@cell_order <- rownames(g$plot$data)[order(g$plot$data$ident)]
 
@@ -317,7 +322,7 @@ setMethod("load_seurat",
 #' @param average_line_color The color of the average profile.
 #' @param standardizing Whether rows should be divided by standard deviation.
 #' @param ceil A value for ceiling (NULL for no ceiling). Ceiling is performed after log transformation, centering and standardization.
-#' @param floor A floor value (NULL for no ceiling). flooring is performed after log transformation, centering and standardization.
+#' @param floor A value for flooring (NULL for no flooring). Flooring is performed after log transformation, centering and standardization.
 #' @param centering Whether rows should be centered. 
 #'
 #' @return A ggplot diagram.
@@ -431,18 +436,18 @@ setMethod(
       if(length(object@cell_types) > 0 & length(object@cell_colors) > 0){
         print_msg("Adding cell populations to the diagram.", msg_type="DEBUG")
         cell_types <- NULL # Avoid "no visible binding for global variable" inn R check.
-        p <- p + ggplot2::geom_vline(aes(xintercept= samples, color=cell_types))
-        p <- p + ggplot2::scale_color_manual(values=object@cell_colors,  guide = ggplot2::guide_legend(override.aes = list(size = 5)))
+        p <- p + geom_vline(aes(xintercept= samples, color=cell_types))
+        p <- p + scale_color_manual(values=object@cell_colors,  guide = guide_legend(override.aes = list(size = 5)))
       }
 
       if(! average_only){
         print_msg("Adding gene profile.", msg_type="DEBUG")
-        p <- p + ggplot2::geom_line(color = "azure3", aes(group = gene), size=0.1)
+        p <- p + geom_line(color = "azure3", aes(group = gene), size=0.1)
       }
 
       print_msg("Adding average profile.", msg_type="DEBUG")
 
-      p <- p + ggplot2::geom_line(
+      p <- p + geom_line(
         data = m_melt %>%
           group_by(cluster, samples) %>%
           summarise(cluster_mean = mean(value)),
@@ -516,9 +521,10 @@ setMethod(
 #' plot_dist
 #' @description
 #' Plot the observed and simulated distance with the Kth nearest neighbors.
-#' @param file_name The name of the input file containing distance and threshold values.
-#' @param path a character string representing the data directory where
-#' input file is stored. Default to current working directory.
+#' @param object A ClusterSet object.
+#' @param path A character string representing the data directory where
+#' input file containing distances and cutting threshold is stored. 
+#' Default to current working directory.
 #'
 #' @return A ggplot diagram.
 #' @export
@@ -529,16 +535,18 @@ setMethod(
 #' @rdname plot_dist
 
 
-plot_dist <-  function(file_name,
+plot_dist <-  function(object,
                        path = ".") {
+  #Extract name of the file containing distance and cutting threshold.
+  file_name <- object@opt_name
   
+  #Path to the file containg distance and cutting threshold.
   if (path == ".") path <- getwd()
-  
   file_path <- file.path(path, file_name)
   file_path <- gsub(pattern = "//", replacement = "/", x = file_path)
+  
+  #Read file containing distance and cutting threshold.
   opt_data <- readLines(file_path)
-  
-  
   
   # Extract cutting threshold value
   dknn <- opt_data[(which(opt_data == ">>thresholds")+1):length(opt_data)]
@@ -599,6 +607,171 @@ plot_dist <-  function(file_name,
   
   return(p)
   
+}
+
+
+
+
+#################################################################
+##    Define the plot_heatmap
+#################################################################
+
+#' @title
+#' plot_heatmap
+#' @description
+#' Plot the observed and simulated distance with the Kth nearest neighbors.
+#' @param object A ClusterSet object.
+#' @param center A logical to indicate whether to center row.. 
+#' @param ceil A value for ceiling (NULL for no ceiling). Ceiling is performed centering.
+#' @param floor A value for flooring (NULL for no flooring). Flooring is performed after centering.
+#' @param cell_order A vector of cell names already ordered.
+#' @param cell_ordering_method The clustering method to be used. This must be "hclust", ADD OTHER CLUSTERING METHOD.
+#' @param use_top_genes A logical to indicate whether to use highly similar genes in the slot top_genes of ClusterSet.
+#' @param name A title for the heatmap.
+#' @param xlab A title for the x axis.
+#' @param ylab A title for the y axis.
+#' @param colorbar_name A title for the colorbar.
+#' @param show_legend A logical to indicate whether to show colorbar.
+#' @param colors A vector of colors.
+#' @param row_labels A logical to indicate whether to show row labels.
+#' @param col_labels A logical to indicate whether to show col labels.
+#' @param label_size A value for label font size.
+#' @param line_size An integer for the horizontal white line size.
+#'
+#' @return Iheatmap-class object
+#' @export
+#'
+#' @examples
+#' m <- matrix(rnorm(6000), nc=20)
+#' m[1:100,1:10] <- m[1:100,1:10] + 4
+#' m[101:200,11:20] <- m[101:200,11:20] + 3
+#' m[201:300,5:15] <- m[201:300,5:15] + -2
+#' 
+#' res <- DBFMCL(data=m,
+#' distance_method="pearson",
+#' av_dot_prod_min = 0,
+#' inflation = 1.2,
+#' k=25,
+#' fdr = 10)
+#' 
+#' plot_heatmap(object = res)
+#' 
+
+#' @rdname plot_heatmap
+
+plot_heatmap <- function(object,
+                         center = TRUE,
+                         ceil = 1,
+                         floor = -1,
+                         cell_order = NULL,
+                         cell_ordering_method = "hclust",
+                         use_top_genes = FALSE,
+                         name = NULL,
+                         xlab = NULL,
+                         ylab = NULL,
+                         colorbar_name = "Expression level",
+                         show_legend = TRUE,
+                         colors = c("#A9D6E5", "#2166AC", "#000000", "#B2182B", "#FFCA3A"),
+                         row_labels = TRUE,
+                         col_labels = FALSE,
+                         label_size = 9,
+                         line_size = 15) {
+  
+  m <- object@data
+  
+  # Centering
+  if(center){
+    print_msg("Centering matrix.", msg_type="DEBUG")
+    m <- t(scale(t(m), center = TRUE, scale = FALSE))
+  }
+  
+  
+  # Cell order
+  if (!is.null(cell_order)){
+    print_msg("Ordering cells.", msg_type="DEBUG")
+    m <- m[,cell_order]
+  } else {
+    if (cell_ordering_method == "hclust"){
+      print_msg("Ordering cells based on hierarchical clustering.", msg_type="DEBUG")
+      m_dist <- as.dist(1 - cor(m, method = 'pearson'))
+      m_clust <- hclust(m_dist, method = 'average')
+      m <- m[,m_clust$order]
+    }
+  }
+  
+  # Ceiling and flooring
+  if(!is.null(ceil)){
+    print_msg("Ceiling matrix.", msg_type="DEBUG")
+    m[m > ceil] <- ceil
+  }
+  
+  if(!is.null(floor)){
+    print_msg("Flooring matrix.", msg_type="DEBUG")
+    m[m < floor] <- floor
+  }  
+  
+  
+  # Add blank row to separate feature clusters in heatmap
+  ## Create blank row
+  blank_row <- data.frame()
+  blank_row[1:line_size, 1:ncol(object@data)] <- NA
+  row.names(blank_row) <- make.names(rep("w_line", line_size), unique=TRUE)
+  colnames(blank_row) <- colnames(object@data)
+  
+  ## Insert blank row in matrix
+  m_blank <- data.frame()
+  for (i in 1:length(object@size)) {
+    row_start <- sum(object@size[0:(i-1)])+1
+    row_end <- sum(object@size[1:i])
+    m_blank_loop <- rbind(m[row_start:row_end,], blank_row)
+    #rownames(test)[(nrow(test)-line_size+1):nrow(test)] <- paste(rep(" ", 2), collapse = '')
+    m_blank <- rbind(m_blank, m_blank_loop)
+  }
+  m <- as.matrix(m_blank)
+  
+  # Reduce m rows to only keep genes from top_genes
+  if(use_top_genes) {
+    if (nrow(object@top_genes) == 1 &
+        ncol(object@top_genes) == 1 &
+        is.na(object@top_genes[1,1])) {
+      stop(paste0("The slot top_genes of the input ClusterSet object is empty. Be sure to run top_genes() before."))
+    }
+    genes_top <- unlist(as.data.frame(t(object@top_genes)), use.names = FALSE)
+    genes_top <- genes_top[!is.na(genes_top)]
+    m <- m[genes_top,]
+  }
+  
+  
+  ####### Heatmap #######
+  # Main heatmap
+  print_msg("Plotting heatmap.", msg_type="DEBUG")
+  
+  htmp <- main_heatmap(data = m,
+                       name = colorbar_name,
+                       show_colorbar = show_legend,
+                       colors = colors,
+                       row_order = seq_len(nrow(m)))
+  
+  htmp <- htmp %>% modify_layout(list(margin = list(t=20, r=10, b=20, l=10)))
+  
+  # Add labels
+  if(row_labels){
+    htmp <- htmp %>% add_row_labels(font = list(size = label_size))}
+  if(col_labels){
+    htmp <- htmp %>% add_col_labels(font = list(size = label_size))}
+  
+  
+  # Add titles
+  if(!is.null(ylab)){
+    htmp <- htmp %>% add_row_title(ylab, side="right", font = list(size = 12))}
+  if(!is.null(xlab)){
+    htmp <- htmp %>% add_col_title(xlab, side="top", font = list(size = 12))}
+  if(!is.null(name)){
+    htmp <- htmp %>% add_col_title(name, side="top", font = list(size = 24))}
+  
+  
+  
+  return(htmp)
 }
 
 
@@ -727,8 +900,7 @@ setMethod(
 #' @param optional_output if TRUE then DBF generate optional output file in the
 #' specified output_path directory. This file contains observed and simulated 
 #' distances, cutting threshold, number of kept genes and FDR value.
-#' @param mcl_cmd_line Boolean. Whether to use the fast MCL version through command line.
-#' @param mcl_cmd_line_threads If mcl_cmd_line is TRUE, how many threads should be used (integer).
+#' @param mcl_threads An integer to determine number of threads for mcl algorithm.
 #' @param distance_method a method to compute the distance to the k-th nearest
 #' neighbor. One of "pearson" (Pearson's correlation coefficient-based
 #' distance), "spearman" (Spearman's rho-based distance), "euclidean".
@@ -805,8 +977,7 @@ DBFMCL <- function(data = NULL,
                    path = ".",
                    output_path = ".",
                    optional_output = TRUE,
-                   mcl_cmd_line=FALSE,
-                   mcl_cmd_line_threads=1,
+                   mcl_threads=1,
                    name = NULL,
                    distance_method = c("pearson", "spearman", "euclidean"),
                    av_dot_prod_min=2,
@@ -818,24 +989,24 @@ DBFMCL <- function(data = NULL,
                    fdr = 10,
                    inflation = 8,
                    set.seed = 123) {
-
+  
   ## testing the system
   if (.Platform$OS.type == "windows") {
     stop("\t--> A unix-like OS is required to launch mcl and cluster programs.")
   }
-
+  
   ## getting parameters
   data_source <- get_data_4_DBFMCL(data = data, filename = filename, path = path)
   data_matrix <- data_source$data
-
+  
   # A simple function to create a random string
   create_rand_str <- function() {
-      v = c(sample(LETTERS, 3, replace = TRUE),
-            sample(0:9, 4, replace = TRUE),
-            sample(letters, 3, replace = TRUE))
-      return(paste0(sample(v),collapse = ""))
+    v = c(sample(LETTERS, 3, replace = TRUE),
+          sample(0:9, 4, replace = TRUE),
+          sample(letters, 3, replace = TRUE))
+    return(paste0(sample(v),collapse = ""))
   }
-
+  
   if (is.null(name)) name <- data_source$name
   if (is.null(name)) name <- create_rand_str()
   
@@ -857,7 +1028,7 @@ DBFMCL <- function(data = NULL,
   txt <- paste("\n\tInflation: ", inflation, sep = "")
   
   ## writting all parameters
-
+  
   cat(
     "The following parameters will be used :",
     "\n\tWorking directory: ", path,
@@ -873,8 +1044,8 @@ DBFMCL <- function(data = NULL,
     "commands: ", silent,
     "\n\tMemory used : ", memory_used, "\n\n"
   )
-
-
+  
+  
   ## DBF algorithm, returns a ClusterSet object
   obj <- DBF(data_matrix,
              output_path = output_path,
@@ -888,115 +1059,92 @@ DBFMCL <- function(data = NULL,
              fdr = fdr,
              set.seed = set.seed
   )
-
+  
   dbf_out_file <- paste0(output_path, "/", name, ".dbf_out.txt")
   dbf_out_file <- gsub(pattern = "//", replacement = "/", x = dbf_out_file)
   mcl_out_file <- paste0(output_path, "/", name, ".mcl_out.txt")
   mcl_out_file <- gsub(pattern = "//", replacement = "/", x = mcl_out_file)
-
+  
   print_msg("DBF completed. Starting MCL step.", msg_type="DEBUG")
-
+  
   if (length(readLines(dbf_out_file)) > 0) {
-
-      ## Launching mcl
-
-      if(mcl_cmd_line){
-        print_msg("Running MCL through the command line for best performance.")
-        mcl_system_cmd(name, inflation = inflation, input_path = output_path, silent = silent, threads = mcl_cmd_line_threads)
+    
+    ## Launching mcl (command line)
+    mcl_system_cmd(name, inflation = inflation, input_path = output_path, silent = silent, threads = mcl_threads)
+    
+    
+    print_msg(paste0("Reading MCL output: ", mcl_out_file), msg_type="DEBUG")
+    
+    ## getting mcl results into the ClusterSet object
+    mcl_cluster <- readLines(mcl_out_file)
+    gene_list <- NULL
+    clusters <- NULL
+    size <- NULL
+    nb <- 0
+    nb_cluster_deleted <- 0
+    
+    for (i in 1:length(mcl_cluster)) {
+      h <- unlist(strsplit(mcl_cluster[i], "\t"))
+      cur_clust <- data_matrix[h,]
+      cur_clust[cur_clust > 0 ] <- 1
+      cur_dot_prod <- cur_clust %*% t(cur_clust)
+      
+      if(mean(cur_dot_prod) > av_dot_prod_min & length(h) > min_cluster_size){
+        
+        nb <- nb + 1
+        gene_list <- c(gene_list, h)
+        clusters <- c(clusters, rep(nb, length(h)))
+        if (is.null(size)) {
+          size <- length(h)
+        }
+        else {
+          size <- c(size, length(h))
+        }
       }else{
-        print_msg("You are using the R implementation of MCL.", msg_type="WARNING")
-        print_msg("Use the command line version for best performance (mcl_cmd_line)", msg_type="WARNING")
-        print_msg(paste("Reading : ", dbf_out_file), msg_type="WARNING")
-        graph_tab <- read.csv(dbf_out_file, sep=" ", header=F)
-        colnames(graph_tab) <- c("source", "dest", "weight")
-        graph_igraph <-  igraph::graph_from_data_frame(graph_tab, directed=F)
-        graph_adj <- igraph::as_adj(graph_igraph, attr='weight')
-        mcl_res <- MCL::mcl(graph_adj,  expansion = 2, inflation = 8, allow1 = TRUE, addLoops = FALSE)
-        cluster_to_genes <- split(rownames(graph_adj), mcl_res$Cluster)
-        print_msg("Writing gene clusters", msg_type="INFO")
-        print_msg(mcl_out_file, msg_type="DEBUG")
-
-        for(i in 1:length(cluster_to_genes)){
-          write.table(paste(cluster_to_genes[[i]], collapse="\t"),
-                      file=mcl_out_file,
-                      append=TRUE,
-                      col.names=FALSE,
-                      row.names=FALSE,
-                      quote=FALSE)
-        }
+        nb_cluster_deleted <- nb_cluster_deleted + 1
       }
-
-      print_msg(paste0("Reading MCL output: ", mcl_out_file), msg_type="DEBUG")
-
-      ## getting mcl results into the ClusterSet object
-      mcl_cluster <- readLines(mcl_out_file)
-      gene_list <- NULL
-      clusters <- NULL
-      size <- NULL
-      nb <- 0
-      nb_cluster_deleted <- 0
-
-      for (i in 1:length(mcl_cluster)) {
-        h <- unlist(strsplit(mcl_cluster[i], "\t"))
-        cur_clust <- data_matrix[h,]
-        cur_clust[cur_clust > 0 ] <- 1
-        cur_dot_prod <- cur_clust %*% t(cur_clust)
-
-        if(mean(cur_dot_prod) > av_dot_prod_min & length(h) > min_cluster_size){
-
-          nb <- nb + 1
-          gene_list <- c(gene_list, h)
-          clusters <- c(clusters, rep(nb, length(h)))
-          if (is.null(size)) {
-            size <- length(h)
-          }
-          else {
-            size <- c(size, length(h))
-          }
-        }else{
-          nb_cluster_deleted <- nb_cluster_deleted + 1
-        }
-      }
-      print_msg(paste(nb, " clusters conserved after MCL partitioning."),
-                msg_type="INFO")
-      print_msg(paste(nb_cluster_deleted,
-                      " clusters filtered out from MCL partitioning (size and mean dot product)."),
-                msg_type="INFO")
-
-
-      ## build ClusterSet object
-      if (nb > 0) {
-        obj@name <- name
-        obj@data <- as.matrix(data_matrix[gene_list, ])
-        obj@cluster <- clusters
-        obj@size <- size
-
-        centers <- matrix(ncol = ncol(data_matrix), nrow = nb)
-        ## calcul of the mean profils
-        for (i in 1:nb) {
-          centers[i, ] <- apply(obj@data[obj@cluster == i, ],
-            2, mean,
-            na.rm = TRUE
-          )
-        }
-        obj@center <- centers
-
-        ## add DBFMCL parameters used to build this object
-        obj@parameters <- list(
-          distance_method = distance_method,
-          k = k,
-          random = random,
-          fdr = fdr,
-          set.seed = set.seed,
-          inflation = inflation
+    }
+    print_msg(paste(nb, " clusters conserved after MCL partitioning."),
+              msg_type="INFO")
+    print_msg(paste(nb_cluster_deleted,
+                    " clusters filtered out from MCL partitioning (size and mean dot product)."),
+              msg_type="INFO")
+    
+    
+    ## build ClusterSet object
+    if (nb > 0) {
+      obj@name <- name
+      obj@data <- as.matrix(data_matrix[gene_list, ])
+      names(clusters) <- rownames(obj@data)
+      obj@cluster <- clusters
+      obj@size <- size
+      
+      centers <- matrix(ncol = ncol(data_matrix), nrow = nb)
+      ## calcul of the mean profils
+      for (i in 1:nb) {
+        centers[i, ] <- apply(obj@data[obj@cluster == i, ],
+                              2, mean,
+                              na.rm = TRUE
         )
       }
+      obj@center <- centers
+      
+      ## add DBFMCL parameters used to build this object
+      obj@parameters <- list(
+        distance_method = distance_method,
+        k = k,
+        random = random,
+        fdr = fdr,
+        set.seed = set.seed,
+        inflation = inflation
+      )
+    }
   }
   else {
     stop("\t--> There is no conserved gene.\n\n")
   }
   return(obj)
-
+  
 }
 
 ###############################################################
@@ -1120,6 +1268,7 @@ DBF <- function(data,
 
       informative <- a$m2[a$m2 != ""]
       if (length(informative) > 0) {
+        obj@opt_name <- "extra_output-dbfAll.txt"
         obj@data <- as.matrix(data[a[[4]] %in% informative, ])
         obj@cluster <- rep(1, nrow(obj@data))
         obj@size <- nrow(obj@data)
@@ -1194,7 +1343,7 @@ mcl_system_cmd <- function(name, inflation = 2.0, input_path = ".", silent = FAL
     ## Testing mcl installation
     if (system("mcl --version | grep 'Stijn van Dongen'", intern = TRUE) > 0) {
       if (!silent) {
-        cat("Running mcl (graph partitioning)... ")
+        cat("Running mcl (graph partitioning)... \n")
         verb <- ""
       }
       else {
@@ -1305,6 +1454,250 @@ get_data_4_DBFMCL <- function(data = NULL, filename = NULL, path = ".") {
 
   return(list(data = data, name = name))
 }
+
+#################################################################
+##    Define top_genes function for ClusterSet object
+#################################################################
+
+#' @title
+#' Best co-expressed genes from each gene cluster
+#' @description
+#' Extract the highly co-expressed genes of each gene cluster.
+#' @param object A \code{ClusterSet} object.
+#' @param top A value for the number of the most similar genes in gene clusters.
+#' @param cluster A vector of gene cluster identity.
+#'
+#' @return 
+#' @export top_genes
+#'
+#' @examples
+#' 
+#' m <- matrix(rnorm(80000), nc=20)
+#' m[1:100,1:10] <- m[1:100,1:10] + 4
+#' m[101:200,11:20] <- m[101:200,11:20] + 3
+#' m[201:300,5:15] <- m[201:300,5:15] + -2
+#' 
+#' res <- DBFMCL(data=m,
+#'               distance_method="pearson",
+#'               av_dot_prod_min = 0,
+#'               inflation = 1.2,
+#'               k=25,
+#'               fdr = 10)
+#'               
+#'res <- top_genes(object = res, top=500, cluster="all")
+#'res@top_genes
+#' 
+#' \dontrun{
+#' ## with an artificial dataset
+#'
+#' m <- matrix(rnorm(80000), nc = 20)
+#' res <- get_data_4_DBFMCL(data=m)
+#' }
+#' 
+
+top_genes <- function(object,
+                      top = 20,
+                      cluster = "all") {
+  
+   if(unique(cluster == "all")) {
+    cluster <- c(1:length(object@size))
+   }
+  
+  # Display a warning message if there is less than n top genes in a gene cluster
+  loop <- 0
+  for (size in object@size[cluster]){
+    loop <- 1 + loop
+    if(top > size) {
+      warning(paste0("Number of top genes is greater than the number of genes in cluster ", loop, ". All genes will be use and order by similarity rank."))
+    }
+  }
+  
+  # Initialization for the for loop
+  clusters <- object@cluster
+  l_cor_means <- list()
+  genes_top <- matrix(ncol = top)
+  
+  # Extract top co-expressed genes for each gene cluster
+  for (i in cluster) {
+    #Extract gene names in cluster i
+    genes <- names(clusters[which(clusters == i)])
+    
+    #Compute distances between genes in cluster i
+    #Use the same distance used by DBFMCL
+    dist_method <- object@parameters$distance_method
+    
+    if (dist_method == "pearson") {
+      cor_genes <- cor(t(object@data[genes,]), method = "pearson")
+      dist <- 1 - cor_genes
+    }
+    
+    if(dist_method == "spearman") {
+      cor_genes <- cor(t(object@data[genes,]), method = "spearman")
+      dist <- 1 - cor_genes
+    }
+    
+    if(dist_method == "euclidean") {
+      dist <- as.matrix(dist(object@data[genes,], method = "euclidean", upper = TRUE, diag = TRUE))
+      diag(dist) <- NA
+    }
+    
+    
+    #Compute mean correlation for each gene in cluster i
+    dist_means <- colMeans(dist, na.rm = TRUE)
+    dist_means <- dist_means[order(dist_means, decreasing = FALSE)]
+    
+    #Extract top genes with the highest correlation mean
+    genes_top <- rbind(genes_top, names(dist_means[1:top]))
+  }
+  
+  # Prepare genes_top matrix
+  if (length(cluster) > 1) {
+    genes_top <- genes_top[2:(length(cluster)+1),]
+  } else {
+    genes_top <- as.matrix(t(genes_top[2:(length(cluster)+1),]))
+  }
+  colnames(genes_top) <- make.names(rep("gene.top", top+1), unique=TRUE)[2:(top+1)]
+  rownames(genes_top) <- make.names(rep("cluster", length(cluster)+1), unique=TRUE)[2:(length(cluster)+1)]
+  
+  # Put genes_top matrix in object@top_genes
+  object@top_genes <- genes_top
+  
+  # Print
+  print_msg(msg = "Results are stored in top_genes slot of ClusterSet object.",
+            msg_type = "INFO")
+  
+  return(object)
+}
+
+#################################################################
+##    Running gprofiler2 to perform enrichment analysis
+#################################################################
+
+#' @title
+#' Perform gene enrichment analysis.
+##' @description
+#' Perform enrichment analysis on all MCL clusters indepentently and store the results in the cluster_annotations slot of the ClusterSet object.
+#' @param object A \code{ClusterSet} object.
+#' @param specie Specie name, as a concatenation of the first letter of the name and the family name, e.g human - hsapien
+#'
+#' @return A \code{ClusterSet} object
+#' @export enrich_analysis
+#'
+#' @examples
+#' 
+#' \dontrun{
+#' ## Assuming myobject is a ClusterSet object with at least 1 cluster.
+#'
+#' gores <- enrich_analysis(myobject)
+#' }
+
+setGeneric("enrich_analysis",
+    function(object,
+            specie="hsapiens") {
+      standardGeneric("enrich_analysis")
+})
+
+#' @rdname enrich_analysis
+setMethod("enrich_analysis",
+    signature(object = "ClusterSet"),
+    function(object,
+            specie="hsapiens") {
+
+      for(cluster in unique(object@cluster)){
+        print(paste0("Enrichment analysis for cluster ", cluster))
+        cluster_name = paste0("Cluster_", cluster)
+        query = rownames(object@data[object@cluster == cluster,])
+        gostres <- gost(query, organism = "hsapiens", ordered_query = FALSE, significant = TRUE, exclude_iea = T)
+        object@cluster_annotations[[cluster]] = list(result = gostres$result, meta = gostres$meta)
+      }
+      return(object)
+  }
+)
+
+#################################################################
+##    Manhattan-like-plot for enrichment analysis on ClusterSet object
+#################################################################
+
+#' @title
+#' Manhattan-like-plot of enrichment analysis results
+##' @description
+#' Retrieve enrichment analysis results from a ClusterSet object and draw a Manhattan-like-plot.
+#' @param object A \code{ClusterSet} object.
+#' @param clusters  A vector of cluster id to plot.
+#' @param verbose Whether or not to print progression in the console.
+#'
+#' @return A \code{ClusterSet} object
+#' @export enrich_viz
+#'
+#' @examples
+#' 
+#' \dontrun{
+#' ## Assuming myobject is a ClusterSet object with at least 1 cluster.
+#'
+#' enrich_viz(myobject)
+#' }
+
+setGeneric("enrich_viz",
+           function(object,
+                    clusters = "all",
+                    verbose = TRUE) {
+             standardGeneric("enrich_viz")
+           })
+
+#' @rdname enrich_viz
+setMethod("enrich_viz",
+          signature(object = "ClusterSet"),
+          function(object,
+                   clusters = "all",
+                   verbose = TRUE) {
+            
+            if (length(clusters) == 1){
+              if (clusters == "all"){
+                clusters <- unique(object@cluster)
+              }
+            }
+            
+            
+            for (cur_cluster in clusters) {
+              # Check if the current cluster id provided exists
+              if (!(cur_cluster %in% unique(object@cluster))) {
+                stop(paste0("Cluster ", cur_cluster, " doesn't exist."))
+              }
+              
+              # Check if there is a result provided by enrich_analysis function for the current cluster
+              if(is.null(object@cluster_annotations[[cur_cluster]]$result)){
+                print_msg(msg_type = "WARNING",
+                          msg = paste0("No functional enrichment analysis results for cluster ", cur_cluster, ".")) #Continue through the next cluster without plotting
+              } else {
+                
+                # Print an informative message to announce plot of the results of the current cluster
+                if (verbose) {
+                  print_msg(msg_type = "INFO",
+                            msg = paste0("Plot enrichment analysis results for cluster ", cur_cluster))
+                }
+                
+                # Create a plotly result plot
+                plot_enrich <- gostplot(object@cluster_annotations[[cur_cluster]],
+                                        interactive = TRUE)
+                plot_enrich <- plot_enrich %>% plotly::layout(title = paste0("Cluster ", cur_cluster),
+                                                              xaxis = list(title = 'Database'))
+                
+                # Store the plot in the cluster_annotation slot
+                object@cluster_annotations[[cur_cluster]]$plot <- plot_enrich
+                print(plot_enrich)
+              }
+            }
+            
+            # Print a message to inform which slot contains the results
+            if (verbose){
+              print_msg(msg_type = "INFO",
+                        msg = paste("Plots are stored in object@cluster_annotations[[<cluster>]]$plot"))
+            }
+            
+            return(object)
+          }
+)
+
 #########################################################
 ##      END PACKAGE scigenex
 #########################################################